backend/requirements.txt
google-generativeai==0.7.2
Flask>=3.0.3
flask-cors==3.0.10
python-dotenv==1.0.1